requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.2.0
python-telegram-bot>=20.0
pyyaml>=6.0
aiohttp>=3.9.4
